+++ src/effect/base.ts
@@ -83,11 +83,9 @@ export class Base implements BaseObject {
    * Apply this effect to a target at the given time
    *
    * @param target
-   * @param reltime - the movie's current time relative to the layer
-   * (will soon be replaced with an instance getter)
    * @abstract
    */
-  apply (target: Movie | Visual, reltime: number): void {} // eslint-disable-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function
+  apply (target: Movie | Visual): void {} // eslint-disable-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function
 
   /**
    * The current time of the target